+++ coursework/include/element/ElementData.hpp
@@ -7,9 +7,9 @@ struct ElementData final {
     sf::Vector2f point;
     float angle;
     ElementType type;
-    int num;
+    size_t num;
     sf::Vector2f scale;
     ElementData(const sf::Vector2f& point, const float angle,
-                const ElementType type, const int num, const sf::Vector2f& scale) :
+                const ElementType type, const size_t num, const sf::Vector2f& scale) :
         point(point), angle(angle), type(type), num(num), scale(scale) { }
 };
