+++ coursework/src/game/entity/component/gun/GunManager.cpp
@@ -4,7 +4,7 @@
 GunManager::GunManager(const EntityCreator &entity_creator, const CollisionManager &collision_manager, QuadtreeEl &quadtree):
     _entity_creator(&entity_creator), _collision_manager(&collision_manager), _quadtree(&quadtree){}
 
-void GunManager::createGun(const sf::Vector2f& p, const int num) {
+void GunManager::createGun(const sf::Vector2f& p, const size_t num) {
     Gun gun(_entity_creator->createGun(p, LEAN, num, _offset_factor));
     _quadtree->insert(&_guns.emplace(gun.getId(), std::move(gun)).first->second.getElement());
 }
